==> 2_build_audio_vectors.ipynb <==
## Build Audio Vectors
Now that the labels have been extracted, we'll use the compiled csv (df_iemocap.csv) to split the original wav files into multiple frames


import librosa
import os
import soundfile as sf
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.style as ms
from tqdm import tqdm
import pickle

import IPython.display
import librosa.display
ms.use('seaborn-v0_8-talk')
%matplotlib inline

y adalah data audio dalam bentuk array satu dimensi yang merepresentasikan amplitudo suara dari file WAV tersebut.
sr adalah angka integer yang menunjukkan berapa banyak sampel audio per detik (sampling rate), di sini 44100 Hz.

file_path = 'data/IEMOCAP_full_release/Session1/dialog/wav/Ses01F_impro01.wav'

y, sr = librosa.load(file_path, sr=44100)
y, sr

## Loop through all the files

import pandas as pd
import math

labels_df = pd.read_csv('data/pre-processed/df_iemocap.csv')
iemocap_dir = 'data/IEMOCAP_full_release/'

The following cells take some time until completely executed

sr = 44100
audio_vectors = {}
# for sess in range(1,6):  # using one session due to memory constraint, can replace [5] with range(1, 6)
for sess in [1]:  # using one session due to memory constraint, can replace [5] with range(1, 6)
    wav_file_path = '{}Session{}/dialog/wav/'.format(iemocap_dir, sess)
    orig_wav_files = os.listdir(wav_file_path)
    for orig_wav_file in tqdm(orig_wav_files):
        try:
            orig_wav_vector, _sr = librosa.load(wav_file_path + orig_wav_file, sr=sr)
            orig_wav_file, file_format = orig_wav_file.split('.')
            for index, row in labels_df[labels_df['wav_file'].str.contains(orig_wav_file)].iterrows():
                start_time, end_time, truncated_wav_file_name, emotion, val, act, dom = row['start_time'], row['end_time'], row['wav_file'], row['emotion'], row['val'], row['act'], row['dom']
                start_frame = math.floor(start_time * sr)
                end_frame = math.floor(end_time * sr)
                truncated_wav_vector = orig_wav_vector[start_frame:end_frame + 1]
                audio_vectors[truncated_wav_file_name] = truncated_wav_vector
        except:
            print('An exception occured for {}'.format(orig_wav_file))
    with open('data/pre-processed/audio_vectors_{}.pkl'.format(sess), 'wb') as f:
        pickle.dump(audio_vectors, f)

Kode ini secara otomatis memotong file audio panjang menjadi segmen-segmen kecil berdasarkan anotasi waktu yang ada di label, lalu menyimpan potongan audio tersebut dalam bentuk array untuk digunakan dalam analisis atau pelatihan model machine learning.
