==> 4_prepare_data.ipynb <==
## Build Speech data files

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler
from IPython.display import display

%matplotlib inline

kode ini merubah dari 7 kelas emosi menjadi 5 kelas emosi

df = pd.read_csv('data/pre-processed/audio_features.csv')
df = df[df['label'].isin([0, 1, 2, 3, 4, 5, 6, 7])]
print(df.shape)
display(df.head())

# change 7 to 2
df['label'] = df['label'].map({0: 0, 1: 1, 2: 1, 3: 2, 4: 2, 5: 3, 6: 4, 7: 5})
df.head()

print(dict(pd.Series(df['label']).value_counts()))

df.to_csv('data/no_sample_df.csv')

# oversample fear
df = pd.concat([df] + [fear_df]*30 + [sur_df]*10, ignore_index=True)




df.to_csv('data/modified_df.csv')
df.head()
df.tail()

print(dict(pd.Series(df['label']).value_counts()))

df.tail()


unique_values = df['label'].unique()
print(unique_values)


import numpy as np

# Ambil 5 data pertama
first_five =  df.head(5)

# Nilai minimal dan maksimal per kolom
min_vals = np.min(first_five, axis=0)
max_vals = np.max(first_five, axis=0)

print("Nilai minimal per kolom pada 5 data pertama:", min_vals)
print("Nilai maksimal per kolom pada 5 data pertama:", max_vals)



unique_values = df['label'].unique()
print(unique_values)


import numpy as np
df = pd.read_csv('data/modified_df.csv')

emotion_dict = {'ang': 0,
                'hap': 1,
                'sad': 2,
                'neu': 3,}

# emotion_dict = {'ang': 0,
#                 'hap': 1,
#                 'exc': 2,
#                 'sad': 3,
#                 'fru': 4,
#                 'fea': 5,
#                 'sur': 6,
#                 'neu': 7,
#                 'xxx': 8,
#                 'oth': 8}


scalar = MinMaxScaler()
df[df.columns[3:]] = scalar.fit_transform(df[df.columns[3:]])
df.head()


unique_values = df['label'].unique()
print(unique_values)

unique_count = df['label'].nunique()
print(f"Jumlah nilai unik di kolom 'nama_kolom': {unique_count}")



x_train, x_test = train_test_split(df, test_size=0.20)

x_train.to_csv('data/s2e/audio_train.csv', index=False)
x_test.to_csv('data/s2e/audio_test.csv', index=False)

print(x_train.shape, x_test.shape)

print(dict(pd.Series(x_train['label']).value_counts()))

## Define preprocessing functions for text

import unicodedata

def unicodeToAscii(s):
    return ''.join(
        c for c in unicodedata.normalize('NFD', s)
        if unicodedata.category(c) != 'Mn'
    )

# Lowercase, trim, and remove non-letter characters
def normalizeString(s):
    s = unicodeToAscii(s.lower().strip())
    s = re.sub(r"([.!?])", r" \1", s)
    s = re.sub(r"[^a-zA-Z.!?]+", r" ", s)
    return s

## Build Text data files

import re
import os
import pickle

useful_regex = re.compile(r'^(\w+)', re.IGNORECASE)

file2transcriptions = {}

for sess in [1]:
    transcripts_path = 'data/IEMOCAP_full_release/Session{}/dialog/transcriptions/'.format(sess)
    transcript_files = os.listdir(transcripts_path)
    for f in transcript_files:
        with open('{}{}'.format(transcripts_path, f), 'r') as f:
            all_lines = f.readlines()

        for l in all_lines:
            audio_code = useful_regex.match(l).group()
            transcription = l.split(':')[-1].strip()
            # assuming that all the keys would be unique and hence no `try`
            file2transcriptions[audio_code] = transcription
# save dict
with open('data/t2e/audiocode2text.pkl', 'wb') as file:
    pickle.dump(file2transcriptions, file)
len(file2transcriptions)

audiocode2text = pickle.load(open('data/t2e/audiocode2text.pkl', 'rb'))

# Prepare text data
text_train = pd.DataFrame()
text_train['wav_file'] = x_train['wav_file']
text_train['label'] = x_train['label']
text_train['transcription'] = [normalizeString(audiocode2text[code]) for code in x_train['wav_file']]

text_test = pd.DataFrame()
text_test['wav_file'] = x_test['wav_file']
text_test['label'] = x_test['label']
text_test['transcription'] = [normalizeString(audiocode2text[code]) for code in x_test['wav_file']]

text_train.to_csv('data/t2e/text_train.csv', index=False)
text_test.to_csv('data/t2e/text_test.csv', index=False)

print(text_train.shape, text_test.shape)
